refactor: restructures storage contract
