+++ src/components/pages/storage/myoffers/StorageEditOfferPage.tsx
@@ -9,7 +9,7 @@ import { AddTxPayload } from 'context/Blockchain/blockchainActions'
 import BlockchainContext from 'context/Blockchain/BlockchainContext'
 import { OfferEditContextProps, StorageBillingPlan, TokenAddressees } from 'context/Market/storage/interfaces'
 import OfferEditContext from 'context/Market/storage/OfferEditContext'
-import StorageContract from 'contracts/Storage'
+import StorageContract from 'contracts/storage/contract'
 import { PeriodInSeconds } from 'models/marketItems/StorageItem'
 import { UIError } from 'models/UIMessage'
 import React, {
@@ -120,7 +120,7 @@ const StorageEditOfferPage: FC<{}> = () => {
     } catch (error) {
       reportError(new UIError({
         error,
-        id: 'contract-storage-set-offer',
+        id: 'contract-storage',
         text: 'Could not set the offer in the contract.',
       }))
       setIsProcessing(false)
