+++ src/context/storage/buy/checkout/Context.tsx
@@ -155,7 +155,7 @@ const Provider: FC = ({ children }) => {
           size,
           token: TokenAddressees[token],
         }
-        const storageContract = (await import('contracts/Storage')).default.getInstance(web3 as Web3)
+        const storageContract = (await import('contracts/storage/contract')).default.getInstance(web3 as Web3)
         appDispatch({
           type: 'SET_IS_LOADING',
           payload: {
