+++ src/contracts/interfaces.ts
@@ -1,7 +1,7 @@
 import { MarketplaceContractErrorId } from './Marketplace'
 import { RifContractErrorId } from './Rif'
 import { RnsContractErrorId } from './Rns'
-import { StorageContractErrorId } from './Storage'
+import { StorageContractErrorId } from './storage/contract'
 
 export type Web3ErrorId = 'web3-getGasPrice'
