+++ src/contracts/storage/interfaces.ts
@@ -1,0 +1,6 @@
+export interface OfferContractData {
+  availableSizeMB: string
+  periods: number[][]
+  prices: string[][]
+  tokens: string[]
+}
